MAT-4393 refactor code
